[Jupyter] Better handling of PFSConentManager errors (#9494)
Wraps the `PFSContentManager.get` call to return the full path to the
content and propagate errors through the MountDrive. Using the full path
to the content allows the frontend to display a more useful error
message. Propagating the errors is more correct and allows a modal to
pop up to display the error message.