Djanicek/allow upload timeout (#9516)
pachhub is currently upgrading. Because of this, requests are going to
pachyderm to store reports for grafana and circle ci is timing out
waiting for them to finish. This step was intended to never fail so it
wouldn't disrupt CI pipelines if GCP was down. I added a timeout so we
never hit the circle-ci timeout and fail.