Update branch deletion logic (#9507)
Notable changes:
- Deleting a branch on which other branches depend for triggers is now
not allowed, unless the user sets `force=true`
- Previous branch deletion logic at the PFS service did a lot of
complicated provenance operations to ensure that downstream branches of
the branch being deleted is aware of this deletion. We no longer need to
do that because in the new relational schema, we only store the edges of
the provenance/subvenance graph, and the whole graph is computed at
read-time.
- Also unified `force-deletion` semantics for both provenance and
triggers. At the end of the day, both provenance and triggers model a
DAG of branches. Deleting a branch should remove it from both DAGs in a
consistent way.