Propagate force delete through PFS APIs (#9541)
We model DAGs in our database. Sometimes we would like to remove nodes
from this DAG. However, sometimes we would like our DAG to prevent us
from accidentally deleting nodes that has dependencies, but sometimes we
don't care and just want to remove the node from the DAG; we use a flag
called `force` to toggle this behaviour. Note, don't confuse `force`
with recursive deletes, we are not doing that here.

This CL makes sure that when the user calls `DeleteAll()`, we use
`force` delete on the repos, which propagates the force to delete all
the branches within the repo. This should help alleviate some of the
recent test flakes we are seeing with `DeleteAll()`.