Refactor the way we watch for events related to commits. (#9465)
This CL is mostly just a refactor of our existing approach to watching
for events related to Commits. There are no new ideas here. I simply
abstracted away the common patterns into a re-usable set of functions,
deleted the repeated code, and used the new stuff instead.

Maybe one new idea here is having separate explicit handlers for
Put/Update events versus Delete events. This is because fundamentally,
delete events don't have commit infos. This is to improve safety, since
without distinguishing the two, the client might accidentally try to
retrieve a non-existent commit info.

TODO: We also need this for projects, repos, and branches. Can we think
of a way to genericize the code further?