remove mount-server from jupyterlab-pachyderm extension (#9563)
This change overhauls our Jupyterlab extension to use the Pachyderm SDK
over the mount-server so that FUSE is no longer needed. This permits
users to run the extension without the use of privileged containers and
reduces the number of moving parts. A full document explaining the
rationale and design can be found
[here](https://www.notion.so/2023-08-25-FUSE-less-JupyterLab-Extension-fa1e77aaa9cb49b394af49e5ada86a29?pvs=4).

---------

Co-authored-by: Kevin Yang <kevin.yang@hpe.com>