Djanicek/core 861/redo test buckets (#9572)
# Summary 

This PR removes the manual test bucketing and the `unit_test` build tag,
and tries to remove the bespoke set up for each type of integration
tests. :tada:


Because merging any of this means it will be immediately used by
developers, it is difficult to break into small pieces without risking
breaking dev workflows. I'm happy to walkthrough the changes in person
though so message me if you'd like to!

# Fixes(see tickets for resolutions)

- https://pachyderm.atlassian.net/browse/CORE-1412
- https://pachyderm.atlassian.net/browse/CORE-1556
- https://pachyderm.atlassian.net/browse/CORE-861
- https://pachyderm.atlassian.net/browse/CORE-1936
- https://pachyderm.atlassian.net/browse/CORE-1875

# Categories of changes

To help with review, you can look at this PR as having a few categories
of changes.
**Most interesting changes:**

- Code to split tests across runners and run `go test` is in
`src/testing/cmds/harness`
- This has a collector and a runner folder. This allows us to collect
the tests while other setup happens in CI(minikube start, docker build)
to help with pipeline performance. The runner reads this and runs the
specified tests distributed to each runner.
- minikubetestenv now relies on k8s leases for locking. See
`client_factory.go` and `deploy.go` .
- This allows us to run tests across multiple processes, with each
process being able to lock.
- There is also a change to add a configMap to each test cluster with a
hash of the helm values. This lets us know if we need to re-install a
cluster, or if we can re-use it without a reinstall.
- `main.yml` for the CI pipeline is changed to use the new runners.

**Important but less interesting changes:**

- `unit_test` tags deleted from the tops of a bunch of files. This is
why the file count is so high.
- Fixing test flakes exposed by having tests bucketed from other,
random, tests from different packages. We have to be alot more hygienic
with our setup/teardown, which helps with flake overall.
- Deleting newly(or previously) dead code.
- Some test reporting code needed changes to handle running packages
independently like in
`src/testing/cmds/go-test-results/collector/main.go`

So for review, I'd start with `src/testing/cmds/harness`, then move to
`deploy.go` then `client_factory.go`, and the rest of the changes are
just indirectly supporting that.

# Performance

We have new knobs to turn to help with the performance of our tests. We
can tune:

- max number of clusters in `clusters.pool`. This is effectively a max
concurrency for integration tests we can use to ensure we don't
provision more k8s namespace than we have cpus for.
- `go test -parallel` this controls the number of parallel tests _within
a package_ that can run. This respects `t.Parallel()` in code.
- `-thread` on `harness/test-runner` controls the number of packages
that run in parallel. Basically, the harness will execute this many `go
test` commands at once.
- number of runners in circle-ci.
- size of runners on circle-ci.

For Performance of integration and unit tests we are generally CPU
bound. The current setup of 9 2xlarge for integration tests and 5 xlarge
for unit tests comes out to roughly the same cost per run as what we
currently run. However, we can change the number and size of runners to
scale to as fast as we want. The down side is it cost more money for
runners.

The configuration with the larger runners means we can re-use clusters
more often and execute faster assuming we can run enough clusters on a
single runner. After testing, our 2Xlarge runners can support 10
concurrent namespaces, up from 4 before this change.

The current configuration runs more tests in roughly the same time as
the previous test bucket setup.

# Follow-ups that are not fixed:

In order to _try_ to keep this scoped I have created tickets for
follow-ups that are not strictly necessary, but would be nice to have.
Let me know if you think any of these are super High priority.
- moving setup from `main.yaml` to Go. Now, all setup is in `main.yaml`
of the Go tests themselves, rather than a bunch of random scripts. It
should all be in Go eventually, but this is a good towards it.
[INFRAENG-289](https://hpe-aiatscale.atlassian.net/browse/INFENG-289)
- Get unit test running with integration tests. I think eventually it
would be nice to have everything runnable from the same command, but
right now, the integration tests seem to clobber the postgres DB that
integration tests use. Once that and any other setup conflicts are
fixed, the test runner should make it trivial to run everything
together.
[INFRAENG-290](https://hpe-aiatscale.atlassian.net/browse/INFENG-290)
- ~~Code coverage is 8% lower and I don't know why.~~ Fixed, needed to
past test.gocoverdir through the raw command.
- Do something better with port-offset. I updated retry logic on the
install, so we shouldn't see test flake from portOffset collisions, but
it's still gross that we could have a port collision. Ticket:
[INFRAENG-283](https://hpe-aiatscale.atlassian.net/browse/INFENG-283)
- ~~TestDatumSetCache [has an error in a goroutine after finishing
sometimes](https://app.circleci.com/pipelines/github/pachyderm/pachyderm/24984/workflows/44e22263-3ea3-4bfe-9d19-20c2f31ca5ae/jobs/442381).
This seems to be pre-existing.~~ It **was** pre-existing, but it turns
out I caused it fixing a separate flake! I think I fixed both now
though.
- TestAutoscalingPipeline continues to be flaky, but is mostly not with
a sleep. It should not be flaky at all, and should not need a sleep.
Ticket: [CORE-2056](https://pachyderm.atlassian.net/browse/CORE-2056)