Djanicek/test yml cleanup (#9588)
For the test bucket change, there was an issue if the test-collector was
running at the same time as a cache restore or go mod download, as that
could change dependency or go binaries as the collector was using for
`go test -list` as it was running. This lead to weird SIGBUS errors.

I put a fix in earlier just to get it
working(https://github.com/pachyderm/pachyderm/pull/9585), but while I
was doing it I noticed a bunch of weirdness with our caching and how
steps were ordered that was slowing down the test jobs. This PR
re-orders our dependency installing and caching to speed the integration
tests back up to where they were before we hit the dependency race
issue.