Bazelify "make proto" (#9578)
CORE-2067

This builds the Go side of protos with `bazel run //:make_proto`. Like
the Docker-based `make proto`, this updates your working copy in-place.
You don't need anything installed other than Bazelisk to run this;
install Bazelisk as `bazel`, and it downloads the right version of
Bazel, and then Bazel makes available all the necessary dependencies for
the actual run.sh script.

The PR as a whole introduces Bazel and sets up a comfortable workspace.

CI continues to run make_proto and fails if there are diffs. This has
been moved from the static-checks PR job to the check-protos job, so we
only need Bazel in one place.

Documentation about the current state is available in `BAZEL.md`.