[INT-1151] Add config file write support (#9600)
Adds the ability to write out config files. This is intended to be an
internal feature.

In order to make this work, I had to do some non-backwards compatible
changes to the `ConfigFile.\_\_init\_\_` method. User's really have no
reason to be touching this code, so I feel comfortable with this change.