[INT-1160] Rename "Mount" to "Load" in Explore and Test views (#9621)
https://pachyderm.atlassian.net/browse/INT-1160

Updated the few instances of `Mount` and `Unmount` to be `Load` and
`Unload` respectively. Each instance mentioned in the story was updated
and I also searched for both strings in the `./jupyter-extension/`
directory. Test ids/findByText references for both strings were also
updated. There are definitely a lot more references to mount/unmount I
noticed in our classes and function names, but those seem out of scope
for the story.

I made a few updates to the README.md to note that python 3.8.X should
be used and `--ip=0.0.0.0` needs to be specified for the Linux VM
development environment. They aren't really related to this story, but
they were such small changes that it seemed unnecessary to create
another PR.