Check If a Chunk Exists After Uploading It (#9634)
The code introduced by this pull request checks if a chunk exists after
attempting to create that chunk in the backend store. This change is
motivated by a few support tickets that were caused by missing chunk
errors. The goal is to hopefully catch these errors close to write time.