Build a docker image as the base for CI runs (#9644)
This builds an image for CircleCI tests. It already has all the
dependencies needed to run bazel installed, so CI runs don't have to do
that every time anymore.

The image includes bazel, but if .bazelversion in the repository
changes, it will fetch that version and not use its cached copy. That
way if nobody feels like rebuilding the container, it's not a big deal,
it will just take a few seconds to download a newer version of bazel.

It's sort of built with bazel; `bazel run //etc/ci-image:push` will
build and push a new version after you edit the Dockerfile. This is how
we bake in the version of bazel mentioned in `.bazelversion`.

A bunch of people asked me to use Nix for this. I tried, but I ended up
with an image that was about 1.5GB and didn't work with circle. (For
those following along at home, I used
nixpkgs.dockerTools.[buildImage|buildNixShellImage], but neither were
particularly great. I also tried `rules_nixpkgs` and added nixpkgs
directly to a Debian base image; that doesn't work because nxpkgs
dynamically links everything, including go binaries like bazelisk, with
a libc in /nix/store/<some hash that depends on the exact version of the
package you get>, and the binaries can't be run on other linuxes.)