Djanicek/infraeng 342/OIDC logic (#9617)
I'm definitely open to discussion on this if we want to remove any of
these changes. I personally haven't seen agood reason to keep things the
way they are but there could definitely be a reason that I am unaware
of.

Teams may occasionally want to deploy pachyderm+determined as enterprise
without enabling authentication. This is a little odd, but it seems like
a reasonable ask for a new enterprise customer to want to try out
enterprise without authentication as they get up and running or do some
initial testing/evaluation.

The determined helm chart allows this configuration, but not the unified
chart. The pachyderm side often just checks that determined is enabled,
and then assumes oidc is enabled. This PR does 3 things to resolve this:

1. Change the pachyderm chart to check `oidc` and `oidc.enabled` instead
of `determined.enabled` for oidc related items like client secret.
2. Change the determined chart in the pachyderm repo to match the
determined chart in the determined repo for the scim and odic defaults.
As they were set(default=true), it was impossible to turn off
`oidc.enabled` if `oidc` was set because helm's templating assumes that
any false-y value (like `false`) is unset, even when manually set. So,
setting `oidc.enabled` to false is read as unset, and so `oidc.enabled`
is always true. This cuts down on maintenance too because we no longer
deviate from the original determined helm chart here.
3. Smoke test was removed because the test was changed. I'll add a
backlog item to re-add a better e2e test for this.