Bosterbuhr/fix det priorityclass CORE-2047 (#9640)
checking to see if a priority class exists before deploying is causing
issue when someone attempts a `helm upgrade`. If the Priority classes do
exist, then they get removed from the kubernetes manifest when
upgrading, and then deleted in the cluster. This removes our old
handling and adds in Determined's newer handling which gives the
deployer manual control over the priority classes.

With this change, the tests also need to be updated to manually handle
running determined in two separate namespaces.

---------

Co-authored-by: Brody Osterbuhr <34581663+BOsterbuhr@users.noreply.github.com>