[INT-1161] Paginated background loading of files for View Datum and PFS (#9665)
Currently if a user attempts to load a directory with a lot of files
(~10000+) from the `view_datum` or `pfs` endpoint that response can take
several seconds causing a noticeable lag which is poor user experience.

This PR changes our custom FileBrowser to load an initial page of 400
files and render that quickly then queues a non blocking promise to load
the rest of the files in 400 file pages. This takes advantage of the
pagination provided by `pfs.client.list_files` by adding a `number` and
`pagination_marker` parameter to the `view_datum` and `pfs` endpoints.

Two integration tests were added for each handler. Also a manual setup
script `setup_large_repo.py` was added for easily setting up a large
repo for testing this behavior locally. I think I might try adding a
cypress test for this, but I can handle it in a follow up PR.

# PFS Pagination Demo

https://github.com/pachyderm/pachyderm/assets/401518/5a2d960d-5276-46c6-ae37-f3486b9f25a2

# View Datum Pagination Demo

https://github.com/pachyderm/pachyderm/assets/401518/b6c38729-3675-4e92-be7f-29542e3d3926