Build a container with pachd built in coverage mode (#9686)
The k8s tests want to collect code coverage from the binary running in
k8s. `rules_go` does not have a way to do that, so I hacked something
together instead. We use `rules_go` to dump the GOPATH of the transitive
closure of pachd into a directory, then we `GO111MODULE=off go build
...` in that directory. A snag is that Bazel runfiles can be symlinks,
but `//go:embed` does not like symlinks. So there is a wrapper script,
`//src/cmd/compile-with-coverage` that resolves all the symlinks and
produces a symlink-free `$GOPATH`. It then invokes `go build` against
that `GOPATH`. The only downside of this approach is that no caching
occurs; if any source file changes, then the entire binary is built from
scratch. That takes about 30s on my 32-core workstation, which is a very
long time.

The rest of the PRs are some genrules to plumb all of that into Bazel,
so that eventually `//oci:pachd_coverage_image` can be produced using
the normal container-building machinery.

Right now, this is somewhat difficult to use.  To run it:

```
$ bazel build //oci:pachd_coverage_image
$ bazel run //src/testing/pachdev -- load-image oci:$(bazel info bazel-bin)/oci/pachd_coverage_image pachd:coverage1234
$ kubectl edit deployment pachd 
<edit to use pachd:coverage1234>
```
Making this simpler to use involves splitting out pachdev's subcommands
into individual bazel targets and then making pachdev invoke the correct
bazel target. That's because I don't want to take a dependency from
pachdev to //oci:pachd_coverage_image, because any change to any code
causes it to be rebuilt from scratch, and that would make every command
slow even when you don't want code coverage. This will all be cleaned up
soon.

We also need to convert the output of `pachctl debug profile cover` to
LCOV and stick it in $COVERAGE_DIRECTORY. Then bazel will merge that
coverage information into everything else that the tests produce,
including the final merged coverage report.

Part of CORE-2081.

---------

Co-authored-by: Robert Uhl <robert.uhl@hpe.com>