[Jupyter] Tabs (lumino) (#9696)
This implements a Tab layout of the extension using some builtin
`lumino` functionality. This shifts us away from a UI model that that
can be summarized as "explore view with a bunch of child views which we
manually show and hide" to "views split between tabs that the user can
'freely' navigate between". To minimize the amount of changes included
in this PR, we still have some situations where we force the user into
certain views, such as:
* when the user is not connected to/logged into a cluster they are
brought to the config screen,
* if a server error happens we show a full screen error.

To accomplish this I had to make the following changes:
* Move the "explore" view into a component,
* Remove all of the showConfig/showDatum/showPipeline/etc. flags and
signals as they are no longer needed,
* Collapsed the Pipeline and PipelineSplash screens into the same react
component,