Upgrade bazel (#9708)
This changes us to Bazel 7.0.2. I don't think it fixes any bugs we care
about, but better to upgrade now than later.

While I was making a new CI image, I added a few quality of life things;
less and an updated terminfo database. The tests actually use less, so
this changes which codepath some pachctl tests use but most people have
less so this is more representative of the real world.