[Jupyter] Allow uninstantiated pachyderm clients (#9705)
Currently when trying to start the extension, if a pachyderm config file
is not present, we create a pachyderm client that connects to
`localhost:30650`. In doing this, the FE is not able to know if the user
entered this config or not. This PR allows an uninstantiated pachyderm
client on startup, and adds checks when accessing any functionality
which requires the pachyderm client. Additionally, the config file is
now passed into the `setup_handlers` method and added to the application
settings.

Additionally to test this functionality I needed a bit more flexibility
in creating the server that we test against. This led to an exploration
into running both client and server asynchronously within the same
thread. Only the new test uses this functionality at this time.