[Jupyter] Prevent a bad config from being stored (#9711)
When handling `PUT /config` the backend was setting the config and then
testing it. Now we test the new config and store it only if the config
is _not_ invalid. Doing so prevents a user from inputting a bad
pachd_address on the config screen and having the UI present it like
it's a valid address. Surprisingly this didn't require and front end
changes.