[Jupyter] Remove dev_server from tests (#9713)
I converted all the tests away from using the dev_server fixture and
removed the fixture.

Additionally:
* I added back in the python tests that were skipped because of the
missing chunk errors. Doing so required a couple of changes to the names
of the resources within the tests.
* I updated the download code a bit to download to the jupyter's
`root_dir` as opposed to the cwd of the server. I think the "correct"
approach anyways, since you can configure which directory the jupyter
server serves files from, but it also has the added benefit of allowing
us to configure this value during the test so that we can write files to
a temp directory. The way this test was written, you needed to delete
the downloaded files before running the test again.
* I also split up the config and pps/publish tests to their own files.