[INT-1162] Infinite scroll in file explorer (#9715)
This PR implements infinite scroll pagination in our custom JuypterLab
FileBrowser. It does so by setting up a scroll event listener that
increments and decrements an index when user scrolls to the upper 20%
and lower 20% respectively of the FileBrowser Contents DOM node. When
this index changes the FileBrowser is re-rendered. Each index represents
an offset of 200 Contents from the start of the overall cached Contents.
Since only a 700 Contents are visible at any time regardless of index
position or total Contents count this solution can scale with a large
number of files (tested with 500k) since the low Contents limit of the
Juypterlab FileBrowser is always respected.

As part of testing this I also made a few changes to improve the
performance of the each re-render of the FileBrowser.
* Cached the initial shallowResponse after navigating to a directory and
reuse that in each call of _get. Making an GET request on each re-render
is unnecessarily slow and is quite noticeable when infinite scroll
causes a re-render.
* Only re-render for infinite scroll not on each paginated request for
Contents. These requests can happen safely in the background now without
needing to update the pagination UI.

One feature currently not implemented is the ability to jump to anywhere
in the Contents list. Since this solution works off an index this should
be possible in a follow up PR. Dynamically changing the scrollbar height
of the Contents DOM node to match the total count of cached Contents
then determining where the user clicks to set the index should
accomplish this. However this is tricky and worth splitting into a
separate story.

### Demo

https://github.com/pachyderm/pachyderm/assets/401518/0ee94933-49be-461f-b5ec-cb144d9a2f06

### Why no library was used for infinite scroll
Most mature infinite scroll libraries (TanStack scroll, react-window,
react-virtualized) require some sort of framework like React or Svelte
and modify the DOM nodes as the user scrolls. This does not work with
the Juypterlab FileBrowser code since even if we mounted a ReactWidget
that output lumino styled `li` those li would not have the same event
handlers setup by the FileBrowser. On top of that the FileBrowser breaks
completely if Contents DOM nodes are modified outside the expected
render loop.