[PFS-174] Lock all tables during a migration transaction (#9722)
This PR updates the migration code to lock all tables during a migration
transaction. This lock operation should be idempotent with existing
calls to`env.LockTables()` across individual version migrations. The
motivation for the PFS team is to prevent background systems from
writing to the postgres tables during a migration. Otherwise, you could
imagine a background process attempting to insert a row into a table as
that table is being migrated. That would lead to unmigrated data or
other weird states.