[INT-1207] Fix local cypress tests for Arch Linux VM Dev Env (#9728)
Local cypress tests do not run in a local Arch linux vm which is our
preferred development environment people will be onboarded into.

I just updated cypress up to the version that adds ARM support. I tried
updating it to the current version, but ended up in a situation where
due to peerDependencies would have had to update most packages in the
repo. Probably worth doing, but maybe in a separate PR. I did have a few
other packages I had to upgrade as a result of upgrading cypress.

I also noticed `mkdirp` is not being used so I removed it.