[INT-1203] Infinite Scroll w/ ScrollBar (#9746)
This PR refactors the infinite scroll to allow the scrollbar to navigate
to any part of the list of contents. While the results are loading if
the user is scrolled all the way to the bottom then the scrollbar is
scrolled up slightly once new results are added. This prevents them from
being stuck at the bottom of the scrollbar with more results left. Since
this only happens at the bottom their scroll position is not disrupted
anywhere else in the list so they can easily look through results.

This changes the re-rendering and updating of index to happen on every
scroll event. The index now also maps directly to an index in the list
of contents. One of the key things that makes this possible is right
before I pushed the last infinite scroll PR I optimized the `_get`
function by returning cached results so re-rendering the FileBrowser is
quite fast now. Only 500 elements are shown at a time in the FileBrowser
so the solution still scales.


The demo is here. Note the cursor distorts in the screen recording,
however it does not look that way on my monitor. Just some weird bug
with QuickTime recording.

https://github.com/pachyderm/pachyderm/assets/401518/29fe82b5-5fb9-4bae-b0c3-bff25782cb35