sync values for upgrades (#9763)
For Determined tests, we needed to set `createNonNamespacedObjects`. My
original fix for this, only worked on initial deploy and I missed
upgrades. For upgrades we also need to keep both options in sync so that
the upgrade can know if it needs to keep the Priority Classes or not.
Installing with priority classes, and then not including them on a
subsequent update breaks the upgrade.