[PFS-204] Fix path range directories (#9770)
This PR fixes an issue with directories when using path range
functionality. If the upper bound of the path range is past the
directory path, but before the first file path in the directory, then
the directory would not get emitted. Fixing this issue was as simple as
removing the upper bound for the fileset iterator in the file source.
This was the only change needed because we mostly fixed this issue with
a prior fix involving directories not emitting when they were the last
path that should be emitted. Removing the upper bound from the fileset
iterator generalizes that solution such that we always extend past the
end of the path range by one file to ensure directories are emitted. The
upper bound is enforced within the callback of the fileset iterator.