internal/miscutil: fix ReadInto returning partial reads during errors (#9774)
`ReadInto` is used to read streams into a buffer. This is used for
reading chunks out of object storage. Chunks have a set maximum size so
they are a natural fit for this pattern.

The bug is in the check for EOF. The `|| n2 == 0` matches cases where
there is an error and n == 0 (probably most errors)
```
errors.Is(err, io.EOF) || n2 == 0
```