Fix regex used on tags for checking whether coverage builds are used
The `=~` shell operator does not support using `\d` to represent the set
of digits (or any other similar backslash+letter classes). You can
confirm by running

    [[ 1 =~ [0-9] ]]
    echo $?
    [[ 1 =~ \d ]]
    echo $?

which should print `0` and then `1`, indicating that the first match
succeeded and the second failed.

These patterns were working fine before, but now we've reached 2.10 and
the second digit needs to be matched.