[Python-SDK] (Backport) Regen protos without ApiBase servicer (#9740) (#9786)
This is to address a bug [INT-1209] where some of 
the generated code for the PJS api was invalid Python
code. However, this code (the ApiBase class) is not, and almost
certainly will not, be used by the client within the SDK, so we do not
need to generate this portion of the code.