upgrade k8s to 1.29 (#9794)
This updates client-go to 1.29.2. kubectl is upgraded to that version in
both circleci and bazel. minikube is updated to the latest version for
circleci. The kubernetes installed by minikube in CI is upgraded to
1.29.2. helm in circleci is updated to the version that `bazel run
//tools/helm` uses. kind is upgraded to v0.22.0 and kindenv now installs
k8s v1.29.2. (I didn't bump the cluster version number to force people
to upgrade; 1.29.0 is probably adequate for testing.)

Upgrading client-go resulted in the pointer package becoming deprecated.
It was replaced with `ptr` and code using `pointer` has been updated to
use `ptr`.

This makes the supported versions of Kubernetes 1.28, 1.29, and 1.30.
1.27 is EOL in June, about when this branch gets released, so this seems
like the appropriate version to support.