[PFS-220] Add upper configuration to prefetcher (#9811)
This PR adds an upper configuration to the fileset prefetcher and
simplifies the fileset source abstraction. A recent change to the
fileset source abstraction removed the upper bound for the underlying
index reader that is used for fileset iteration. The reason for this
change was to handle an edge case where a directory had files that were
not contained within the configured path range. The fileset prefetcher
depended on this upper bound for bounding its prefetching. In this PR,
we added an upper bound configuration such that the prefetcher can be
explicitly told when it should stop prefetching. Also, the fileset
source abstraction had some redundancy in the index configuration code,
so that was simplified.