Update node version in main.yml (#9797)
Small little thing I noticed. When I updated the node version in .nvmrc
and other parts of build process in a previous PR, I forgot to update
these two places. In an ideal world we would have some single source for
this value, but might as well make this consistent in the short term.