adjust console release logic (#9789)
On the recent releases, the build bot managed to create a release draft
which is great. But since it is a draft, it didn't trigger a release
docker build push which is why it failed. So now I have moved that logic
over to the actual release job, by publishing the release, we will
trigger the docker build push. That way, the `create release draft` job
is truly a test before releasing everything together at the end like it
should.

---------

Co-authored-by: Danny Zhu <dzhu@hpe.com>