Add project metadata (#9836)
This adds a metadata column to the projects table, and lets Go code read
and write it. Something I noticed is that update semantics are pretty
unclear as there was previously only one thing to update. So we make
some effort to make the `CreateProject` RPC not overwrite metadata,
rather than requiring the user to download the project and copy the
metadata into the update request. (This means that `pachctl update
project foo --description bar` doesn't delete foo's metadata.) However,
it seems like the pfsdb functions can safely expect you to read before
you write, so `UpdateProject({})` will clear the metadata and clear the
description.

While I was in pfsdb.projects, I made the update tests check that the
update worked. Previously we only checked that update didn't return an
error, now we check that that the update took effect.

This PR also sets up 2.10.0 to have database migrations.