[INT-1231] Remove unnecessary height (#9837)
So I was not actually able to reproduce this error in either chrome or
firefox. However the CSS shown in the reproduction gif `height: 100%`
that was causing the error can be safely removed so I went ahead and did
that. https://pachyderm.atlassian.net/browse/INT-1231

Here is an image showing it is possible to scroll the publish view.
<img width="385" alt="Screenshot 2024-03-14 at 8 29 21 AM"
src="https://github.com/pachyderm/pachyderm/assets/401518/fb260d82-ce82-499f-90c0-bc6bbd75e450">