Add ibazel (#9841)
This adds [ibazel](https://github.com/bazelbuild/bazel-watcher), which
is bazel that reruns its command whenever the dependencies change.
Documentation has been added to explain how to use it.

The PR also upgrades Bazel to the latest version.

This PR is the first binary that needs to be installed, so I wrote a
bazel rule and Go helper to install a binary to a "good" location. This
lets you run `bazel run //tools/ibazel:install` to install it somewhere
in your $PATH, so that it works every time even if `~/go/bin` isn't on
your path.

I have tried it for tests and pushing to k8s and it works well enough. I
personally wouldn't use it for pushing to k8s after every code edit, but
if you want to, you can. I used it to develop `//src/cmd/install-tool`
with `ibazel test //src/cmd/install-tool:all` and it worked well.

It monitors the job output to look for Gazelle errors and will re-run it
for you, saving you a tiny bit of time!