Djanicek/infraneg 295/first test (#9741)
This PR is for INFENG-295. It is an MVP of a test to validate randomized
pipeline DAGs before and after an upgrade.

This test takes a list of proto files+methods and traverses the fields
in each message, generating data based on the field name if specified,
or the proto data type if no name-specific field is assigned. There are
shenanigans around `Input` since that field is important for generating
valid DAGs and also has some server-side constraints that that the test
needs to know. I'm concerned about listing data-generators on a
per-field basis as it breaks if we change field names, but also that
should only happen when we break the API contract so hopefully that is
uncommon.

It currently works to generate non-trivial dags but has some issues(I
don't love the handling for pipeline inputs) and some not-so-great
pieces(a list field names isn't great) that need to be improved on.

I'd like to get this basic version running, and see how valuable it is
before I sink anymore time into those however. The plan is to let it run
for a while in the pipeline and see if it's useful. If we find it useful
I can circle back and make further improvements. If not, I can just
remove it.

It does generate some pretty cool test data though:
```
$ pachctl draw pipeline Fkqko2Xh_mo_EigIduN --project=z
                    +-----------+             +-----------+             +-----------+             +-----------+
                    |   <A|YD   |             |   %;F<!   |             |    <'i    |             |   ;HDCj   |
                    +-----------+             +-----------+             +-----------+             +-----------+
                         /                         /                           \                       /|                         
                        /                         /                             \                     / |                         
                       |                     -----                ---------------+--------------------  |                         
                       /                    /                    /                \                     |                         
                      /                    /                    /                  \                    |                         
               +-----------+        +-----------+        +-----------+        +-----------+        +-----------+
               |JDvFt_2Ux..|        |cCh_-H2wg..|        |Fkqko2Xh_..|        |i-sss2FbT..|        |    wFh    |
               +-----------+        +-----------+        +-----------+        +-----------+        +-----------+
```