Add //tools/loki and a testloki library to test against a real Loki instance (#9867)
This adds some tools: `//tools/loki` and `//tools/logcli` (loki's client
CLI). We use the tools to implement a `testloki` library, which starts
loki and lets you add data or read data from it. Should make any
loki-related tests easier to write.

Fun fact: loki sleeps and does nothing for 15s when it starts up. There
are config options to adjust the timing; I tried making it 1ms, but that
breaks it. 3s was the lowest that actually worked. I don't know if that
depends on my machine or not, however.